First release
This program allow you to update your CCleaner without having to go manually on the website.
Simply run it and press Update button.

How it work ?
It goes on the CCleaner download page and it fetch sur download url by using the "download" button. Then it use the download URL to download the latest version and install it in silent mode.